Refactor: 인증 서비스와 관리자 세션 경계 정리
