+++ app/src/main/java/com/personal/happygallery/app/admin/port/AdminSession.java
@@ -1,0 +1,6 @@
+package com.personal.happygallery.app.admin.port;
+
+import java.time.Instant;
+
+/** 관리자 인증 세션 정보. */
+public record AdminSession(Long adminUserId, String username, Instant createdAt) {}
